Rename files and bug fixes
Rename main class to BusynessManager
Rename root package to BusynessManager
Correct README.md and AboutUs.md
