+++ docs/team/rozaliesmit.md
@@ -1,0 +1,6 @@
+# Rozalie Smit - Project Portfolio Page
+
+## Overview
+
+
+### Summary of Contributions
